Edition suppression stage

See merge request thimothee_lepetz/gestion-formation!15

+++ database/migrations/2023_04_19_081607_create_salaries_table.php
@@ -48,7 +48,7 @@ return new class extends Migration
             $table->string('rpps')->nullable();
             $table->string('adeli')->nullable();
             $table->string('cpn');
-            $table->integer('taux_emploi');
+            $table->double('taux_emploi');
             $table->double('horaire_contrat');
             $table->double('montant_aq004');
             $table->double('taux_horaire');
